feat(rust): move and refactor entity
